Optimized front-end display
In the former version, I make a md5 encript befor submmit the user's password, but the user will see that the length of password be changed in a moment. This commit optimized the problem.
But it may be no so elegant, I am not profession on front end.
I would be very grateful if you can help to optimize the code.